Modernize sessions redux
