--- src/redux/constants.ts
@@ -1,15 +1,0 @@
-// sessions
-export const FETCH_SESSIONS = 'app/Sessions/FETCH_SESSIONS';
-export const FETCH_SESSIONS_FAILURE = 'app/Sessions/FETCH_SESSIONS_FAILURE';
-export const FETCH_SESSIONS_SUCCESS = 'app/Sessions/FETCH_SESSIONS_SUCCESS';
-export const UPDATE_SESSIONS = 'app/Sessions/UPDATE_SESSIONS';
-
-export const FETCH_USER_FEATURED_SESSIONS = 'app/Sessions/FETCH_USER_FEATURED_SESSIONS';
-export const FETCH_USER_FEATURED_SESSIONS_FAILURE =
-  'app/Sessions/FETCH_USER_FEATURED_SESSIONS_FAILURE';
-export const FETCH_USER_FEATURED_SESSIONS_SUCCESS =
-  'app/Sessions/FETCH_USER_FEATURED_SESSIONS_SUCCESS';
-
-export const SET_USER_FEATURED_SESSIONS = 'app/Sessions/SET_USER_FEATURED_SESSIONS';
-export const SET_USER_FEATURED_SESSIONS_FAILURE = 'app/Sessions/SET_USER_FEATURED_SESSIONS_FAILURE';
-export const SET_USER_FEATURED_SESSIONS_SUCCESS = 'app/Sessions/SET_USER_FEATURED_SESSIONS_SUCCESS';
