+++ src/redux/sessions/state.ts
@@ -1,0 +1,11 @@
+import { SessionsState } from './types';
+
+export const initialSessionsState: SessionsState = {
+  fetching: false,
+  fetchingError: null,
+  list: [],
+  obj: {},
+  featured: {},
+  featuredFetching: false,
+  featuredError: null,
+};
